SmartHChatbot/requirements.txt
# Application Streamlit
streamlit>=1.32.0

# Microsoft Bot Framework (pour Teams)
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
aiohttp>=3.9.0
flask>=3.0.0

# Requêtes HTTP pour Ollama
requests>=2.31.0

# Génération de PDF (optionnel)
reportlab>=4.1.0
fpdf2>=2.7.8

# Utilitaires
python-dotenv>=1.0.1
plotly==5.18.0